+++ order-service/src/main/java/org/dei/order/clients/StoreHouseClient.java
@@ -1,0 +1,9 @@
+package org.dei.order.clients;
+
+import org.dei.storehouse.controller.StoreHouseController;
+import org.springframework.cloud.openfeign.FeignClient;
+
+@FeignClient(url = "localhost:8888")
+public abstract class StoreHouseClient implements StoreHouseController {
+
+}
