+++ payment-service/src/main/java/org/dei/payment/controller/PaymentController.java
@@ -1,33 +1,12 @@
 package org.dei.payment.controller;
 
-import org.dei.payment.converter.OrderConverter;
-import org.dei.payment.converter.UserDetailsConvert;
 import org.dei.payment.dto.OrderDTO;
 import org.dei.payment.dto.UserDetailsDTO;
-import org.dei.payment.service.PaymentService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/api/payment")
-public class PaymentController {
-    private final Logger LOGGER = LoggerFactory.getLogger(PaymentController.class);
-    private final PaymentService paymentService;
-    private final OrderConverter orderConverter;
-    private final UserDetailsConvert userDetailsConvert;
-    @Autowired
-    public PaymentController(PaymentService paymentService, OrderConverter orderConverter, UserDetailsConvert userDetailsConvert) {
-        this.paymentService = paymentService;
-        this.orderConverter = orderConverter;
-        this.userDetailsConvert = userDetailsConvert;
-    }
-
+public interface PaymentController {
     @RequestMapping(path = "/{order_id}", method = RequestMethod.PUT)
-    public OrderDTO perform(@PathVariable("order_id") int orderId, @RequestBody UserDetailsDTO userDetailsDTO) {
-        LOGGER.info("start performing payement");
-        return orderConverter.orderToDTO(paymentService.perform(orderId, userDetailsConvert.DTOtoUserDetails(userDetailsDTO)).getOrder());
-    }
-
+    OrderDTO perform(@PathVariable("order_id") int orderId, @RequestBody UserDetailsDTO userDetailsDTO);
 }
